refactor as general command
